Re #2407 Move BuildConfig out of Stack.Types.Config
