+++ src/Stack/Init.hs
@@ -47,8 +47,9 @@ import           Stack.Runners
                    ( ShouldReexec (..), withConfig, withGlobalProject )
 import           Stack.SourceMap
                    ( SnapshotCandidate, loadProjectSnapshotCandidate )
-import           Stack.Types.Config ( HasConfig, HasGHCVariant, Project (..) )
+import           Stack.Types.Config ( HasConfig, Project (..) )
 import           Stack.Types.Config.Exception ( ConfigPrettyException (..) )
+import           Stack.Types.GHCVariant ( HasGHCVariant )
 import           Stack.Types.GlobalOpts ( GlobalOpts (..) )
 import           Stack.Types.Runner (Runner, globalOptsL )
 import           Stack.Types.Resolver ( AbstractResolver, Snapshots (..) )
