+++ src/Stack/PackageFile.hs
@@ -27,8 +27,8 @@ import           Stack.Constants
                    )
 import           Stack.Constants.Config ( distDirFromDir )
 import           Stack.Prelude
+import           Stack.Types.BuildConfig ( HasBuildConfig (..) )
 import           Stack.Types.CompilerPaths ( cabalVersionL )
-import           Stack.Types.Config ( HasBuildConfig (..) )
 import           Stack.Types.EnvConfig ( HasEnvConfig )
 import           Stack.Types.NamedComponent ( NamedComponent (..) )
 import           Stack.Types.PackageFile
