+++ src/Stack/Types/Config.hs
@@ -14,29 +14,18 @@
 module Stack.Types.Config
   (
   -- * Main configuration types and classes
-  -- ** HasPlatform & HasStackRoot
-    HasPlatform (..)
   -- ** Config & HasConfig
-  , Config (..)
+    Config (..)
   , HasConfig (..)
   , askLatestSnapshotUrl
   , configProjectRoot
   -- ** BuildConfig & HasBuildConfig
-  , BuildConfig (..)
   , ProjectPackage (..)
   , DepPackage (..)
   , ppRoot
   , ppVersion
   , ppComponents
   , ppGPD
-  , stackYamlL
-  , projectRootL
-  , HasBuildConfig (..)
-  -- ** Storage databases
-  , UserStorage (..)
-  , ProjectStorage (..)
-  -- ** GHCVariant & HasGHCVariant
-  , HasGHCVariant (..)
   -- * Details
   -- ** EnvSettings
   , EnvSettings (..)
@@ -56,7 +45,6 @@ module Stack.Types.Config
   -- * Paths
   , bindirSuffix
   , GlobalInfoSource (..)
-  , getProjectWorkDir
   , docDirSuffix
   , platformOnlyRelDir
   , useShaPathOnWindows
@@ -67,7 +55,6 @@ module Stack.Types.Config
   -- * Command-related types
   , module X
   -- * Lens helpers
-  , wantedCompilerVersionL
   , ExtraDirs (..)
   , buildOptsL
   , globalOptsL
@@ -107,10 +94,8 @@ import qualified Distribution.PackageDescription as C
 import           Distribution.System ( Platform )
 import qualified Distribution.Text ( display )
 import           Generics.Deriving.Monoid ( mappenddefault, memptydefault )
-import           Lens.Micro ( _1, _2 )
 import           Options.Applicative ( ReadM )
 import qualified Options.Applicative.Types as OA
-import           Pantry.Internal ( Storage )
 import           Path
                    ( (</>), parent, parseAbsDir, parseAbsFile, parseRelDir
                    , parseRelFile, reldir, relfile
@@ -129,21 +114,20 @@ import           Stack.Types.ConfigMonoid
                    ( ConfigMonoid (..), parseConfigMonoidObject)
 import           Stack.Types.Docker ( DockerOpts )
 import           Stack.Types.DumpLogs ( DumpLogs )
-import           Stack.Types.GHCVariant ( GHCVariant (..) )
+import           Stack.Types.GHCVariant ( GHCVariant (..), HasGHCVariant (..) )
 import           Stack.Types.GlobalOpts ( GlobalOpts (..) )
 import           Stack.Types.NamedComponent ( NamedComponent (..) )
 import           Stack.Types.Nix ( NixOpts )
-import           Stack.Types.PlatformVariant
-                   ( PlatformVariant, platformVariantSuffix )
+import           Stack.Types.Platform
+                   ( HasPlatform (..), PlatformVariant, platformVariantSuffix )
 import           Stack.Types.PvpBounds ( PvpBounds )
 import           Stack.Types.Resolver ( AbstractResolver )
 import           Stack.Types.Runner ( HasRunner (..), Runner, globalOptsL )
 import           Stack.Types.SCM ( SCM )
 import           Stack.Types.SetupInfo ( SetupInfo )
 import           Stack.Types.SourceMap
-                   ( CommonPackage (..), DepPackage (..), ProjectPackage (..)
-                   , SMWanted (..)
-                   )
+                   ( CommonPackage (..), DepPackage (..), ProjectPackage (..) )
+import           Stack.Types.Storage ( UserStorage )
 import           Stack.Types.TemplateName ( TemplateName )
 import           Stack.Types.Version ( VersionCheck (..), VersionRange )
 
@@ -277,23 +261,6 @@ data Config = Config
     -- ^ Turn on Stack developer mode for additional messages?
   }
 
--- | Class for environment values which have a GHCVariant
-class HasGHCVariant env where
-  ghcVariantL :: SimpleGetter env GHCVariant
-  default ghcVariantL :: HasConfig env => SimpleGetter env GHCVariant
-  ghcVariantL = configL.ghcVariantL
-  {-# INLINE ghcVariantL #-}
-
--- | A bit of type safety to ensure we're talking to the right database.
-newtype UserStorage = UserStorage
-  { unUserStorage :: Storage
-  }
-
--- | A bit of type safety to ensure we're talking to the right database.
-newtype ProjectStorage = ProjectStorage
-  { unProjectStorage :: Storage
-  }
-
 -- | The project root directory, if in a project.
 configProjectRoot :: Config -> Maybe (Path Abs Dir)
 configProjectRoot c =
@@ -335,34 +302,6 @@ defaultLogLevel = LevelInfo
 
 readStyles :: ReadM StylesUpdate
 readStyles = parseStylesUpdateFromString <$> OA.readerAsk
-
--- | A superset of 'Config' adding information on how to build code. The reason
--- for this breakdown is because we will need some of the information from
--- 'Config' in order to determine the values here.
---
--- These are the components which know nothing about local configuration.
-data BuildConfig = BuildConfig
-  { bcConfig     :: !Config
-  , bcSMWanted :: !SMWanted
-  , bcExtraPackageDBs :: ![Path Abs Dir]
-    -- ^ Extra package databases
-  , bcStackYaml  :: !(Path Abs File)
-    -- ^ Location of the stack.yaml file.
-    --
-    -- Note: if the STACK_YAML environment variable is used, this may be
-    -- different from projectRootL </> "stack.yaml" if a different file
-    -- name is used.
-  , bcProjectStorage :: !ProjectStorage
-  -- ^ Database connection pool for project Stack database
-  , bcCurator :: !(Maybe Curator)
-  }
-
-stackYamlL :: HasBuildConfig env => Lens' env (Path Abs File)
-stackYamlL = buildConfigL.lens bcStackYaml (\x y -> x { bcStackYaml = y })
-
--- | Directory containing the project's stack.yaml file
-projectRootL :: HasBuildConfig env => Getting r env (Path Abs Dir)
-projectRootL = stackYamlL.to parent
 
 ppGPD :: MonadIO m => ProjectPackage -> m GenericPackageDescription
 ppGPD = liftIO . cpGPD . ppCommon
@@ -488,13 +427,6 @@ ghcInstallHook :: HasConfig env => RIO env (Path Abs File)
 ghcInstallHook = do
   hd <- hooksDir
   pure (hd </> [relfile|ghc-install.sh|])
-
--- | Per-project work dir
-getProjectWorkDir :: (HasBuildConfig env, MonadReader env m) => m (Path Abs Dir)
-getProjectWorkDir = do
-  root    <- view projectRootL
-  workDir <- view workDirL
-  pure (root </> workDir)
 
 -- | Relative directory for the platform identifier
 platformOnlyRelDir ::
@@ -630,17 +562,6 @@ parseProjectAndConfigMonoid rootDir =
 -- Lens classes
 -----------------------------------
 
--- | Class for environment values which have a Platform
-class HasPlatform env where
-  platformL :: Lens' env Platform
-  default platformL :: HasConfig env => Lens' env Platform
-  platformL = configL.platformL
-  {-# INLINE platformL #-}
-  platformVariantL :: Lens' env PlatformVariant
-  default platformVariantL :: HasConfig env => Lens' env PlatformVariant
-  platformVariantL = configL.platformVariantL
-  {-# INLINE platformVariantL #-}
-
 -- | Class for environment values that can provide a 'Config'.
 class ( HasPlatform env
       , HasGHCVariant env
@@ -650,83 +571,39 @@ class ( HasPlatform env
       , HasRunner env
       ) => HasConfig env where
   configL :: Lens' env Config
-  default configL :: HasBuildConfig env => Lens' env Config
-  configL = buildConfigL.lens bcConfig (\x y -> x { bcConfig = y })
-  {-# INLINE configL #-}
-
-class HasConfig env => HasBuildConfig env where
-  buildConfigL :: Lens' env BuildConfig
 
 -----------------------------------
 -- Lens instances
 -----------------------------------
-
-instance HasPlatform (Platform, PlatformVariant) where
-  platformL = _1
-  platformVariantL = _2
 
 instance HasPlatform Config where
   platformL = lens configPlatform (\x y -> x { configPlatform = y })
   platformVariantL =
     lens configPlatformVariant (\x y -> x { configPlatformVariant = y })
 
-instance HasPlatform BuildConfig
-
-instance HasGHCVariant GHCVariant where
-  ghcVariantL = id
-  {-# INLINE ghcVariantL #-}
-
 instance HasGHCVariant Config where
   ghcVariantL = to $ fromMaybe GHCStandard . configGHCVariant
-
-instance HasGHCVariant BuildConfig
 
 instance HasProcessContext Config where
   processContextL = runnerL.processContextL
 
-instance HasProcessContext BuildConfig where
-  processContextL = configL.processContextL
-
 instance HasPantryConfig Config where
   pantryConfigL = lens configPantryConfig (\x y -> x { configPantryConfig = y })
-
-instance HasPantryConfig BuildConfig where
-  pantryConfigL = configL.pantryConfigL
 
 instance HasConfig Config where
   configL = id
   {-# INLINE configL #-}
 
-instance HasConfig BuildConfig where
-  configL = lens bcConfig (\x y -> x { bcConfig = y })
-
-instance HasBuildConfig BuildConfig where
-  buildConfigL = id
-  {-# INLINE buildConfigL #-}
-
 instance HasRunner Config where
   runnerL = lens configRunner (\x y -> x { configRunner = y })
 
-instance HasRunner BuildConfig where
-  runnerL = configL.runnerL
-
 instance HasLogFunc Config where
-  logFuncL = runnerL.logFuncL
-
-instance HasLogFunc BuildConfig where
   logFuncL = runnerL.logFuncL
 
 instance HasStylesUpdate Config where
   stylesUpdateL = runnerL.stylesUpdateL
 
-instance HasStylesUpdate BuildConfig where
-  stylesUpdateL = runnerL.stylesUpdateL
-
 instance HasTerm Config where
-  useColorL = runnerL.useColorL
-  termWidthL = runnerL.termWidthL
-
-instance HasTerm BuildConfig where
   useColorL = runnerL.useColorL
   termWidthL = runnerL.termWidthL
 
@@ -740,11 +617,6 @@ stackRootL = configL.lens configStackRoot (\x y -> x { configStackRoot = y })
 stackGlobalConfigL :: HasConfig s => Lens' s (Path Abs File)
 stackGlobalConfigL =
   configL.lens configUserConfigPath (\x y -> x { configUserConfigPath = y })
-
--- | The compiler specified by the @SnapshotDef@. This may be different from the
--- actual compiler used!
-wantedCompilerVersionL :: HasBuildConfig s => Getting r s WantedCompiler
-wantedCompilerVersionL = buildConfigL.to (smwCompiler . bcSMWanted)
 
 data ExtraDirs = ExtraDirs
   { edBins :: ![Path Abs Dir]
